refactor: merge ServiceLifecycleCoordinator into ServiceRegistry
Linux kernel/module/main.c handles both symbol table and lifecycle in one
module. Nexus split these across ServiceRegistry (core/) and
ServiceLifecycleCoordinator (system_services/lifecycle/) — a legacy
artifact, not intentional design. This PR merges them.

- Extend ServiceRegistry with all SLC methods: enlist(), swap_service(),
  mark_bootstrapped(), start/stop_persistent_services(),
  activate/deactivate_hot_swappable_services(), classify_all()
- Add dispatch param to ServiceRegistry constructor (kernel passes it)
- Add set_lifecycle_manager() for factory to inject BLM at link-time
- service_coordinator property now returns self._service_registry
- Factory no longer creates SLC separately
- Delete service_lifecycle_coordinator.py (674 lines)
- Update all importers (production + test)
