+++ src/nexus/core/nexus_fs.py
@@ -212,7 +212,7 @@ class NexusFS(  # type: ignore[misc]
         # Populated by factory via enlist_wired_services() at link().
         from nexus.core.service_registry import ServiceRegistry
 
-        self._service_registry: ServiceRegistry = ServiceRegistry()
+        self._service_registry: ServiceRegistry = ServiceRegistry(dispatch=self._dispatch)
 
         # Lifecycle state — set by link() / initialize() / bootstrap()
         self._linked: bool = False
@@ -324,19 +324,13 @@ class NexusFS(  # type: ignore[misc]
     # Registered by factory via enlist_wired_services() at link().
 
     @property
-    def service_coordinator(self) -> Any | None:
-        """ServiceLifecycleCoordinator (set by factory at initialize time)."""
-        return getattr(self, "_service_coordinator", None)
+    def service_coordinator(self) -> Any:
+        """ServiceRegistry with integrated lifecycle (formerly ServiceLifecycleCoordinator)."""
+        return self._service_registry
 
     async def swap_service(self, name: str, new_instance: Any, **kwargs: Any) -> None:
-        """Hot-swap a service: atomic replace → drain → hook swap.
-
-        Requires system services to be available (coordinator must exist).
-        """
-        coord = self.service_coordinator
-        if coord is None:
-            raise RuntimeError("Service hot-swap requires system services (no coordinator)")
-        await coord.swap_service(name, new_instance, **kwargs)
+        """Hot-swap a service: atomic replace → drain → hook swap."""
+        await self._service_registry.swap_service(name, new_instance, **kwargs)
 
     @property
     def namespace_manager(self) -> Any | None:
