+++ tests/unit/server/test_rpc_parity.py
@@ -112,7 +112,7 @@ def test_all_public_methods_are_exposed_or_excluded():
         "initialize",  # Boot phase 2 - one-time side effects, not an RPC operation
         "bootstrap",  # Boot phase 3 - async task startup, server-only
         "service",  # ServiceRegistry lookup — local kernel API, not an RPC operation (Issue #1452)
-        "service_coordinator",  # ServiceLifecycleCoordinator access — server-only (Issue #1452 Phase 3)
+        "service_coordinator",  # ServiceRegistry lifecycle access — server-only (Issue #1452 Phase 3)
         "swap_service",  # Hot-swap via coordinator — server-only admin operation (Issue #1452 Phase 3)
         "load_all_saved_mounts",  # Internal initialization method - called automatically on startup
         # Server-side only methods (clients get this via HTTP headers)
